Updateing the CMakeLists.txt to resolve linking conflict and included new tests
When trying to run GitHub actions, a linking error was found that was not present on my local machine.
After sufficiently clearing various CMake caches, the issue was replicated and it appears to be because of doubly including various cpp files.

Solution: The culprit of the problems is source/geometry.cpp which seems to include various other cpp files. I have excluded this file from the CMakeLists.txt. In the future I suspect this file will be removed completely.

New tests are also included.