Added getEigenvalues method and updated dependencies
In the DiracOperator.cpp and DiracOperator.hpp files, the getEigenvalues method was extracted from the header file and moved to the cpp file for better code organization and encapsulation. This function returns the eigenvalues of the Dirac Operator.

Also, the library 'libhdf5-serial-dev' was added to the list of dependencies in .github/workflows/compatability_tests.yml and .github/workflows/build_and_test.yml to ensure successful compilation and testing across different platforms. This library provides the necessary tools for working with hdf5 data files, which are used in this application for data storage and manipulation.

The build and test stages were also renamed in the compatibility tests workflow for clarity and consistency. This renaming helps to avoid confusion when viewing the results in GitHub Actions.