Add EigenvalueRecorder and refactor HDF5-related options
Added the EigenvalueRecorder class to record and save eigenvalues. This class takes a "DiracOperator" object, a coupling constant "g2", and a "simulationId" string to record the eigenvalues from the Dirac operator into a HDF5 file labelled by the simulation id. It was created to provide a more effective and organized way to handle and record eigenvalues for different simulations.

Furthermore, the definition of ARMA_USE_HDF5 macro was moved from the global CMakeLists.txt file to the EigenvalueRecorder.hpp so that it's only defined where actually needed. This was done to avoid potential conflicts with other parts of the code where HDF5 is not required.

Lastly, the EigenvalueRecorder files were added into the CMakeLists.txt file to ensure they are included in the build.

Files added:
- examples/Type13/EigenvalueRecorder.hpp
- examples/Type13/EigenvalueRecorder.cpp

Files changed:
- CMakeLists.txt
- examples/Type13/CMakeLists.txt