Improving simulation class (#6)
* Adding ability to retrieve a ref to dirac operator from simulation class.

This change is the start of making this library actually useful for calculations.

* Add an eigenvalue recorder for DiracOperator

A new method was added to DiracOperator to compute its eigenvalues. Additionally, a Recorder class was created in the Type13Metropolis example to log these eigenvalues into the chosen output file. This enhancement will aid in debugging, performance monitoring, and further scientific analyses during the simulation process.

* Add HDF5 support and save eigenvalues to file

Added Armadillo's HDF5 support in both 'CMakeLists.txt' and 'Type13Metropolis.cpp'. 'Type13Metropolis.cpp' now includes a 'Recorder' class which can save simulation data to file in HDF5 format. This change aids data storage and analysis as HDF5 allows for easy recall of larger datasets. Also updated CMake files to find and link with HDF5 libraries.

* Added getEigenvalues method and updated dependencies

In the DiracOperator.cpp and DiracOperator.hpp files, the getEigenvalues method was extracted from the header file and moved to the cpp file for better code organization and encapsulation. This function returns the eigenvalues of the Dirac Operator.

Also, the library 'libhdf5-serial-dev' was added to the list of dependencies in .github/workflows/compatability_tests.yml and .github/workflows/build_and_test.yml to ensure successful compilation and testing across different platforms. This library provides the necessary tools for working with hdf5 data files, which are used in this application for data storage and manipulation.

The build and test stages were also renamed in the compatibility tests workflow for clarity and consistency. This renaming helps to avoid confusion when viewing the results in GitHub Actions.

* Add EigenvalueRecorder and refactor HDF5-related options

Added the EigenvalueRecorder class to record and save eigenvalues. This class takes a "DiracOperator" object, a coupling constant "g2", and a "simulationId" string to record the eigenvalues from the Dirac operator into a HDF5 file labelled by the simulation id. It was created to provide a more effective and organized way to handle and record eigenvalues for different simulations.

Furthermore, the definition of ARMA_USE_HDF5 macro was moved from the global CMakeLists.txt file to the EigenvalueRecorder.hpp so that it's only defined where actually needed. This was done to avoid potential conflicts with other parts of the code where HDF5 is not required.

Lastly, the EigenvalueRecorder files were added into the CMakeLists.txt file to ensure they are included in the build.

* Refactor Type13Metropolis.cpp, update build dependencies, and correct a comment typo

Refactored Type13Metropolis.cpp file to simplify the code. A large portion of hand-written HDF5 I/O code was replaced with the more streamlined EigenvalueRecorder class, which reduced the length of the file significantly.

In the build_and_test.yml file, a long line of dependency installation commands was broken down into multiple lines for better readability and easier maintenance.

A minor typo in Simulation.hpp header file comment was corrected for improved code documentation clarity.

These changes were made to make the codebase more maintainable and the comments more accurate.

* Add documentation to EigenvalueRecorder class

Added detailed documentation to EigenvalueRecorder class to enhance code readability and understanding. The comments provide a clear understanding of the purpose and functionality of EigenvalueRecorder class and its recordEigenvalues method. These improvements aim to ease the process of code maintenance and modification.

* Add eigenvalue test and update DiracOperator comments

Added a unit test for DiracOperator to verify that the eigenvalues are being calculated correctly.

* "Refactor Simulation testing and update DiracOperator and Hamiltonian classes

The Simulation unit test has been updated to create a temporary variable `dirac` of a Dirac operator, and validate not only its type but also if the Dirac matrix is correct.

The `Hamiltonian` class was improved by replacing the description, adding more comments to empty `TODO` areas, and replacing dynamic allocation of `double` arrays with `std::vectors`. This change provides safer memory management and prevents memory leaks.

The `DiracOperator` class got a copy constructor. This constructor makes a deep copy of all aspects of the operator, ensuring accurate copying of state details.

Memory allocation for storing previous configuration was improved by replacing a dynamic array `mat_bk` with a vector, leading to cleaner and memory safe code.

Furthermore, the `Hamiltonian` class got updated with safer memory handling by using `std::vector<double>` instead of raw pointers. Raw pointers were deleted later in the code causing potential memory leaks.

The Clifford-type was moved to private data members of DiracOperator for encapsulation.

Added the task in TODO file for making an interface for DiracOperator"

* Tidying github actions files