Using Please.build as build tool.  (#10)
After many months of bashing my head against this, I've finally cracked
it.

I've got both RFL libraries - new and old - being built with Please. 
I've also go all the tests running with Please. 

Please is a platform independent, language independent build tool. It
has out of the box support for Python, Go, C++, Dockerfiles and K8s
config etc.
See the info here: [Please.build](https://please.build). 
It seems to be the internal build tool for a company called
Thought-Machine, which is drastically smaller than Google, so it might
be lacking in some features that Bazel. However, I like it. It is
extensible, and I think that will prove useful.
Its also smart enough to know when to rebuild a project and when it's
happy to use the cache. If something starts going awry, just run `plz
clean`

To build this whole project you can just run `./pleasew build`, and
everything should work (assuming you have Armadillo and GSL installed).
To run all the tests you can just run `./pleasew test` and it should run
all the tests in this project.

If you plan to work on this repo in any way. I'd recommend you install
Please directly, in which case the commands would be `plz build`, `plz
test`.

I've made some minor modifications to the import paths in the new_RFL
project. And I've moved everything around which is probably going to
destroy the diff.

---
Some extra information for posteritiy


I spent a lot of time trying to figure out how dynamic libraries work
with Please, I've just come to the understanding that they not handled
in a special way. Please is about dependencies which are linked and
built in to each other. Runtime loading is not really handled by this
tool. Nor should it, really.

I also spend some time seeing if I could use the github repo pieces of
Please to build the Armadillo and GSL libraries. You can - but I've not
included the build targets for that.